fix: Redirect back to creators if not authenticated (QA 82)
